+++ derive/src/derive/core/mod.rs
@@ -1,0 +1,10 @@
+/// `core::convert`
+pub(crate) mod convert;
+/// `core::fmt`
+pub(crate) mod fmt;
+/// `core::future`
+pub(crate) mod future;
+/// `core::iter`
+pub(crate) mod iter;
+/// `core::ops`
+pub(crate) mod ops;