+++ derive/src/derive/mod.rs
@@ -1,3 +1,5 @@
+pub(crate) mod core;
 pub(crate) mod external;
+#[cfg(feature = "std")]
 pub(crate) mod std;
 pub(crate) mod ty_impls;